feat: replace std::indirect with constexpr Box, use static_vector for args
Replace Box<A> (std::indirect) with a constexpr-capable owning pointer
that supports nullable default construction, enabling static_vector
storage. Change comp_apply.args from std::vector to static_vector.
Add static_assert tests proving full compile-time evaluation of the
Mendler interpreter pipeline (read → elaborate → core_to_comp → eval).

+++ src/smd/smdscheme/sender/fixpoint_eval.hpp
@@ -129,8 +129,7 @@ mendler_run(Comp<MaxList> const &comp,
                                 std::get<comp_lambda<CompT, MaxList>>(
                                     lam_layer);
 
-                            if (app.args.size() !=
-                                static_cast<size_t>(lam.params.size()))
+                            if (app.args.size() != lam.params.size())
                                 return Res{foundation::parse_error{
                                     {}, "arity mismatch"}};
 
@@ -141,7 +140,7 @@ mendler_run(Comp<MaxList> const &comp,
                                                : closure::env<CompT, 16>{env};
 
                             // Evaluate args in CURRENT env, bind in NEW env
-                            for (size_t i = 0; i < app.args.size(); ++i) {
+                            for (int i = 0; i < app.args.size(); ++i) {
                                 auto arg_r = mendler_run<MaxList, MaxBindings>(
                                     *app.args[i], env);
                                 if (!arg_r.has_value())
